refactor: organize exceptions by domain
