+++ api/src/main/kotlin/siksha/wafflestudio/api/common/CrawlerApiKeyFilter.kt
@@ -9,7 +9,7 @@ import org.springframework.http.HttpMethod
 import org.springframework.stereotype.Component
 import org.springframework.web.filter.OncePerRequestFilter
 import org.springframework.web.servlet.HandlerExceptionResolver
-import siksha.wafflestudio.core.domain.common.exception.UnauthorizedUserException
+import siksha.wafflestudio.core.domain.common.exception.auth.UnauthorizedUserException
 
 @Component
 class CrawlerApiKeyFilter(
