+++ api/src/main/kotlin/siksha/wafflestudio/api/controller/v1/PostController.kt
@@ -21,7 +21,7 @@ import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import siksha.wafflestudio.api.common.userId
-import siksha.wafflestudio.core.domain.common.exception.InvalidPostFormException
+import siksha.wafflestudio.core.domain.common.exception.community.InvalidPostFormException
 import siksha.wafflestudio.core.domain.v1.community.post.dto.CreatePostReportRequestDto
 import siksha.wafflestudio.core.domain.v1.community.post.dto.PaginatedPostsResponseDto
 import siksha.wafflestudio.core.domain.v1.community.post.dto.PostCreateRequestDto
