+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/auth/service/AuthService.kt
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service
 import siksha.wafflestudio.core.domain.auth.JwtProvider
 import siksha.wafflestudio.core.domain.auth.dto.AuthResponseDto
 import siksha.wafflestudio.core.domain.auth.social.data.SocialProfile
-import siksha.wafflestudio.core.domain.common.exception.UnauthorizedUserException
+import siksha.wafflestudio.core.domain.common.exception.auth.UnauthorizedUserException
 import siksha.wafflestudio.core.domain.user.repository.UserRepository
 import siksha.wafflestudio.core.domain.user.service.UserService
 import java.time.Duration
