+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/common/exception/ExceptionHandler.kt
@@ -6,6 +6,8 @@ import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestControllerAdvice
+import siksha.wafflestudio.core.domain.common.exception.auth.UnauthorizedUserException
+import siksha.wafflestudio.core.domain.common.exception.user.UserNotFoundException
 
 @RestControllerAdvice
 class ExceptionHandler {
