+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/common/exception/main/MainException.kt
@@ -1,0 +1,11 @@
+package siksha.wafflestudio.core.domain.common.exception.main
+
+import org.springframework.http.HttpStatus
+import siksha.wafflestudio.core.domain.common.exception.SikshaException
+
+sealed class MainException(
+    httpStatus: HttpStatus,
+    errorMessage: String,
+) : SikshaException(httpStatus, errorMessage)
+
+class RestaurantNotFound : MainException(HttpStatus.NOT_FOUND, "Restaurant not found")
