+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/common/exception/main/RestaurantException.kt
@@ -1,6 +1,7 @@
-package siksha.wafflestudio.core.domain.common.exception
+package siksha.wafflestudio.core.domain.common.exception.main
 
 import org.springframework.http.HttpStatus
+import siksha.wafflestudio.core.domain.common.exception.SikshaException
 
 sealed class RestaurantException(
     httpStatus: HttpStatus,
