+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/main/menu/service/MenuV2Service.kt
@@ -4,12 +4,12 @@ import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import siksha.wafflestudio.core.domain.common.exception.InvalidMealTypeException
-import siksha.wafflestudio.core.domain.common.exception.MenuAlarmAlreadyExistsException
-import siksha.wafflestudio.core.domain.common.exception.MenuAlarmException
-import siksha.wafflestudio.core.domain.common.exception.MenuNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.MenuNotLikedException
-import siksha.wafflestudio.core.domain.common.exception.UserNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.InvalidMealTypeException
+import siksha.wafflestudio.core.domain.common.exception.main.MenuAlarmAlreadyExistsException
+import siksha.wafflestudio.core.domain.common.exception.main.MenuAlarmException
+import siksha.wafflestudio.core.domain.common.exception.main.MenuNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.MenuNotLikedException
+import siksha.wafflestudio.core.domain.common.exception.user.UserNotFoundException
 import siksha.wafflestudio.core.domain.main.meal.data.MealType
 import siksha.wafflestudio.core.domain.main.meal.repository.MealMenuV2Repository
 import siksha.wafflestudio.core.domain.main.menu.dto.MenuV2AlarmDto
