+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/main/review/service/ReviewV2Service.kt
@@ -6,15 +6,15 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.multipart.MultipartFile
-import siksha.wafflestudio.core.domain.common.exception.KeywordMissingException
-import siksha.wafflestudio.core.domain.common.exception.MenuNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.NotReviewOwnerException
-import siksha.wafflestudio.core.domain.common.exception.ReviewAlreadyExistsException
-import siksha.wafflestudio.core.domain.common.exception.ReviewAndMenuMismatchException
-import siksha.wafflestudio.core.domain.common.exception.ReviewNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.ReviewSaveFailedException
-import siksha.wafflestudio.core.domain.common.exception.SelfReviewLikeNotAllowedException
-import siksha.wafflestudio.core.domain.common.exception.UserNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.KeywordMissingException
+import siksha.wafflestudio.core.domain.common.exception.main.MenuNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.NotReviewOwnerException
+import siksha.wafflestudio.core.domain.common.exception.main.ReviewAlreadyExistsException
+import siksha.wafflestudio.core.domain.common.exception.main.ReviewAndMenuMismatchException
+import siksha.wafflestudio.core.domain.common.exception.main.ReviewNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.ReviewSaveFailedException
+import siksha.wafflestudio.core.domain.common.exception.main.SelfReviewLikeNotAllowedException
+import siksha.wafflestudio.core.domain.common.exception.user.UserNotFoundException
 import siksha.wafflestudio.core.domain.image.data.Image
 import siksha.wafflestudio.core.domain.image.data.ImageCategory
 import siksha.wafflestudio.core.domain.image.repository.ImageRepository
