+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/v1/community/comment/service/CommentService.kt
@@ -5,17 +5,17 @@ import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import siksha.wafflestudio.core.domain.common.exception.CommentAlreadyReportedException
-import siksha.wafflestudio.core.domain.common.exception.CommentNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.CommentReportSaveFailedException
-import siksha.wafflestudio.core.domain.common.exception.CustomNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.InvalidCommentReportFormException
-import siksha.wafflestudio.core.domain.common.exception.InvalidPageNumberException
-import siksha.wafflestudio.core.domain.common.exception.NotCommentOwnerException
-import siksha.wafflestudio.core.domain.common.exception.NotFoundItem
-import siksha.wafflestudio.core.domain.common.exception.PostNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.UnauthorizedUserException
-import siksha.wafflestudio.core.domain.common.exception.UserNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.CommentAlreadyReportedException
+import siksha.wafflestudio.core.domain.common.exception.community.CommentNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.CommentReportSaveFailedException
+import siksha.wafflestudio.core.domain.common.exception.community.CustomNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.InvalidCommentReportFormException
+import siksha.wafflestudio.core.domain.common.exception.community.InvalidPageNumberException
+import siksha.wafflestudio.core.domain.common.exception.community.NotCommentOwnerException
+import siksha.wafflestudio.core.domain.common.exception.community.NotFoundItem
+import siksha.wafflestudio.core.domain.common.exception.community.PostNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.auth.UnauthorizedUserException
+import siksha.wafflestudio.core.domain.common.exception.user.UserNotFoundException
 import siksha.wafflestudio.core.domain.user.repository.UserRepository
 import siksha.wafflestudio.core.domain.v1.community.comment.data.Comment
 import siksha.wafflestudio.core.domain.v1.community.comment.data.CommentLike
