+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/v1/community/post/service/PostService.kt
@@ -8,16 +8,16 @@ import org.springframework.data.domain.PageRequest
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
-import siksha.wafflestudio.core.domain.common.exception.BoardNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.CustomNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.InvalidPageNumberException
-import siksha.wafflestudio.core.domain.common.exception.InvalidPostReportFormException
-import siksha.wafflestudio.core.domain.common.exception.NotFoundItem
-import siksha.wafflestudio.core.domain.common.exception.NotPostOwnerException
-import siksha.wafflestudio.core.domain.common.exception.PostAlreadyReportedException
-import siksha.wafflestudio.core.domain.common.exception.PostNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.PostReportSaveFailedException
-import siksha.wafflestudio.core.domain.common.exception.UnauthorizedUserException
+import siksha.wafflestudio.core.domain.common.exception.community.BoardNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.CustomNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.InvalidPageNumberException
+import siksha.wafflestudio.core.domain.common.exception.community.InvalidPostReportFormException
+import siksha.wafflestudio.core.domain.common.exception.community.NotFoundItem
+import siksha.wafflestudio.core.domain.common.exception.community.NotPostOwnerException
+import siksha.wafflestudio.core.domain.common.exception.community.PostAlreadyReportedException
+import siksha.wafflestudio.core.domain.common.exception.community.PostNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.PostReportSaveFailedException
+import siksha.wafflestudio.core.domain.common.exception.auth.UnauthorizedUserException
 import siksha.wafflestudio.core.domain.image.data.Image
 import siksha.wafflestudio.core.domain.image.data.ImageCategory
 import siksha.wafflestudio.core.domain.image.repository.ImageRepository
