+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/v1/main/restaurant/service/RestaurantService.kt
@@ -3,8 +3,8 @@ package siksha.wafflestudio.core.domain.v1.main.restaurant.service
 import jakarta.transaction.Transactional
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.stereotype.Service
-import siksha.wafflestudio.core.domain.common.exception.RestaurantNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.UserNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.RestaurantNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.user.UserNotFoundException
 import siksha.wafflestudio.core.domain.user.repository.UserRepository
 import siksha.wafflestudio.core.domain.v1.main.restaurant.data.RestaurantCustom
 import siksha.wafflestudio.core.domain.v1.main.restaurant.dto.RestaurantLikeResponseDto
