+++ core/src/test/kotlin/service/post/PostServiceTest.kt
@@ -13,12 +13,12 @@ import org.springframework.data.domain.PageImpl
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.http.HttpStatus
-import siksha.wafflestudio.core.domain.common.exception.BoardNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.InvalidPostReportFormException
-import siksha.wafflestudio.core.domain.common.exception.NotPostOwnerException
-import siksha.wafflestudio.core.domain.common.exception.PostAlreadyReportedException
-import siksha.wafflestudio.core.domain.common.exception.PostNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.UnauthorizedUserException
+import siksha.wafflestudio.core.domain.common.exception.community.BoardNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.InvalidPostReportFormException
+import siksha.wafflestudio.core.domain.common.exception.community.NotPostOwnerException
+import siksha.wafflestudio.core.domain.common.exception.community.PostAlreadyReportedException
+import siksha.wafflestudio.core.domain.common.exception.community.PostNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.auth.UnauthorizedUserException
 import siksha.wafflestudio.core.domain.image.repository.ImageRepository
 import siksha.wafflestudio.core.domain.user.data.User
 import siksha.wafflestudio.core.domain.user.repository.UserRepository
