+++ core/src/test/kotlin/service/restaurant/CustomV2ServiceTest.kt
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import siksha.wafflestudio.core.domain.common.exception.InvalidCustomException
+import siksha.wafflestudio.core.domain.common.exception.main.InvalidCustomException
 import siksha.wafflestudio.core.domain.main.restaurant.data.BuildingV2
 import siksha.wafflestudio.core.domain.main.restaurant.data.CustomV2
 import siksha.wafflestudio.core.domain.main.restaurant.data.RestaurantV2
