+++ core/src/test/kotlin/service/review/ReviewServiceTest.kt
@@ -21,14 +21,14 @@ import org.mockito.kotlin.whenever
 import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Sort
-import siksha.wafflestudio.core.domain.common.exception.CommentNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.InvalidScoreException
-import siksha.wafflestudio.core.domain.common.exception.KeywordMissingException
-import siksha.wafflestudio.core.domain.common.exception.MenuNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.ReviewAlreadyExistsException
-import siksha.wafflestudio.core.domain.common.exception.ReviewNotFoundException
-import siksha.wafflestudio.core.domain.common.exception.SelfReviewLikeNotAllowedException
-import siksha.wafflestudio.core.domain.common.exception.UserNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.community.CommentNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.InvalidScoreException
+import siksha.wafflestudio.core.domain.common.exception.main.KeywordMissingException
+import siksha.wafflestudio.core.domain.common.exception.main.MenuNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.ReviewAlreadyExistsException
+import siksha.wafflestudio.core.domain.common.exception.main.ReviewNotFoundException
+import siksha.wafflestudio.core.domain.common.exception.main.SelfReviewLikeNotAllowedException
+import siksha.wafflestudio.core.domain.common.exception.user.UserNotFoundException
 import siksha.wafflestudio.core.domain.image.repository.ImageRepository
 import siksha.wafflestudio.core.domain.user.data.User
 import siksha.wafflestudio.core.domain.user.repository.UserRepository
